Replaces hard-coded input file name with parameter - Closes #4
The program LifeGrid now accepts a string parameter which it resolves as
the name of an input file.  If no parameter is given, it will default to
glider.txt (which is currently malformed, parses as 3 columns wide).

The dimensions of the grid held in the input file are also no longer hard-
coded, but measured by opening the file before the method that reads the
data.